requirements.txt
fastapi
uvicorn[standard]
pdfplumber
PyMuPDF
pytesseract
reportlab
faker
pillow
python-multipart
requests
transformers
torch
sqlalchemy
streamlit
pytest
sqlalchemy
python-dotenv
python-dateutil
